Extract BufferLengthControl from a private nested class to a public class
